seperate out useMarketImage
