+++ pages/markets/[marketid].tsx
@@ -1,10 +1,15 @@
 import { Disclosure, Transition } from "@headlessui/react";
-import { FullMarketFragment, MarketStatus } from "@zeitgeistpm/indexer";
+import {
+  FullMarketFragment,
+  MarketStatus,
+  ScoringRule,
+} from "@zeitgeistpm/indexer";
 import {
   MarketOutcomeAssetId,
   ScalarRangeType,
   parseAssetId,
 } from "@zeitgeistpm/sdk";
+import LatestTrades from "components/front-page/LatestTrades";
 import { MarketLiquiditySection } from "components/liquidity/MarketLiquiditySection";
 import DisputeResult from "components/markets/DisputeResult";
 import { AddressDetails } from "components/markets/MarketAddresses";
@@ -23,10 +28,16 @@ import CategoricalReportBox from "components/outcomes/CategoricalReportBox";
 import ScalarDisputeBox from "components/outcomes/ScalarDisputeBox";
 import ScalarReportBox from "components/outcomes/ScalarReportBox";
 import Amm2TradeForm from "components/trade-form/Amm2TradeForm";
+import { TradeTabType } from "components/trade-form/TradeTab";
+import ReferendumSummary from "components/ui/ReferendumSummary";
 import Skeleton from "components/ui/Skeleton";
 import { ChartSeries } from "components/ui/TimeSeriesChart";
 import Decimal from "decimal.js";
 import { GraphQLClient } from "graphql-request";
+import {
+  CmsMarketMetadata,
+  getCmsMarketMetadataForMarket,
+} from "lib/cms/get-market-metadata";
 import { PromotedMarket } from "lib/cms/get-promoted-markets";
 import { ZTG, environment, graphQlEndpoint } from "lib/constants";
 import {
@@ -57,21 +68,13 @@ import { parseAssetIdString } from "lib/util/parse-asset-id";
 import { NextPage } from "next";
 import dynamic from "next/dynamic";
 import Image from "next/image";
+import Link from "next/link";
 import { useRouter } from "next/router";
 import NotFoundPage from "pages/404";
 import { useEffect, useMemo, useState } from "react";
 import { AlertTriangle, ChevronDown, X } from "react-feather";
 import { AiOutlineFileAdd } from "react-icons/ai";
 import { FaChevronUp } from "react-icons/fa";
-import { ScoringRule } from "@zeitgeistpm/indexer";
-import { TradeTabType } from "components/trade-form/TradeTab";
-import ReferendumSummary from "components/ui/ReferendumSummary";
-import {
-  CmsMarketMetadata,
-  getCmsMarketMetadataForMarket,
-} from "lib/cms/get-market-metadata";
-import LatestTrades from "components/front-page/LatestTrades";
-import Link from "next/link";
 
 const TradeForm = dynamic(() => import("../../components/trade-form"), {
   ssr: false,
